Move `useTermsSuggestions` to its own folder
